server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
itsdangerous==2.1.2
Werkzeug==2.3.7
Jinja2==3.1.2
email-validator==2.0.0
cryptography==41.0.3
eventlet==0.33.3
gunicorn==21.2.0
SQLAlchemy==2.0.20
alembic==1.12.0
WTForms==3.0.1
urllib3==2.0.4
six==1.16.0
MarkupSafe==2.1.3
idna==3.4
bidict==0.22.1
python-engineio==4.7.1
python-socketio==5.9.0